+++ src/neptune_lightgbm/impl/__init__.py
@@ -46,9 +46,7 @@ except ImportError:
         verify_type,
     )
 
-from neptune_lightgbm._version import get_versions
-
-__version__ = get_versions()["version"]
+from neptune_lightgbm.impl.version import __version__
 
 INTEGRATION_VERSION_KEY = "source_code/integrations/neptune-lightgbm"