+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-## [UNRELEASED] neptune-lightgbm 0.10.0
+## neptune-lightgbm 0.10.0
 
 ### Changes
-
 - Moved `neptune-lightgbm` package to `src` directory ([#8](https://github.com/neptune-ai/neptune-lightgbm/pull/8))
 - Defaulted `log_trees_as_dataframe` to `false` ([#5](https://github.com/neptune-ai/neptune-lightgbm/pull/5))
 - Updated `create_booster_summary` to log tree dataframe as a csv ([#5](https://github.com/neptune-ai/neptune-lightgbm/pull/5))
+- Poetry as a package builder ([#?](?))
 
 ## Fixes